Added a couple more rules for matching icons
We can now fallback to Bluetooth and USB icons. They're not the best
icons, and they will often not be visible due to low score, but still
better than the generic icon of an internal card.

Also added `alsa-input-mic` as one of the possible matches for the
microphone. This will likely make all microphones look the same. Sorry.